chore(starknet_integration_tests): separate utils, use proper fn
commit-id:04e90fc2
